The checks for busy SPI TX have now been placed just before each TX to the ST7735.
In this way all the graphic functions can give back control to the mcu
immediately while the DMA transfer is still executing, resulting in more
time for the mcu to do other non-graphic stuff.

I modified the DemoTFT code to run as a benchmark without pauses. The
demo running time in portrait mode went from 1850ms to 1713ms.
I expect an even bigger speed increase in a multitasking environment
(FreeRTOS)